Added sendVerify
